% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo1.R
\name{ELBO1}
\alias{ELBO1}
\title{ELBO function for DPMM with varied alpha and fixed sigma}
\usage{
ELBO1(D, N, C0, T0, s1, s2, X, L20, Mu0, W1, W2, V1, V2, L1, L2, Plog)
}
\arguments{
\item{D}{Number of Dimensions}

\item{N}{Number of Observations}

\item{C0}{Covariance matrix of the Observed data}

\item{T0}{Number of clusters in the variational distribution}

\item{s1}{Shape parameter of the prior Gamma distribution for alpha}

\item{s2}{Rate parameter of the prior Gamma distribution for alpha}

\item{X}{Observations, N x D matrix}

\item{L20}{Precision parameter of the prior multivariate normal distribution for eta}

\item{Mu0}{Mean parameter of the prior multivariate normal distribution for eta}

\item{W1}{Shape parameter of the posterior gamma distribution for alpha}

\item{W2}{Rate parameter of the posterior gamma distribution for alpha}

\item{V1}{1st shape parameters of the posterior beta distribution for v's, T0 x 1 matrix}

\item{V2}{2nd shape parameters of the posterior beta distribution for v's, T0 x 1 matrix}

\item{L1}{Mean parameters of the posterior multivariate normal distribution for eta's, T0 x D matrix}

\item{L2}{Precision parameters of the posterior multivariate normal distribution for eta's, T0 x 1 matrix}

\item{Plog}{Posterior probabilities of the latent allocations}
}
\value{
the value of the ELBO
}
\description{
This function computes the ELBO for a DPMM with varied alpha and fixed sigma
}
\examples{
ELBO1(D=2,N=100,C0=diag(2),T0=20,s1=0.001,s2=0.001,X=matrix(1,nrow=100, ncol=2),L20=0.001,Mu0=matrix(c(rep(0,2)), nrow=1),W1=0.001,W2=0.001,V1=matrix(0.0001, nrow = 20, ncol = 1),V2=matrix(0.0001, nrow = 20, ncol = 1),L1=matrix(0.0001, nrow = 20, ncol = 2),L2=matrix(0.0001, nrow = 20, ncol = 1),Plog=matrix(-3,nrow=100,ncol=20))
}
