% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfvi1.R
\name{mfvi1}
\alias{mfvi1}
\title{Mean-Field Variational Inference-I}
\usage{
mfvi1(
  d,
  n,
  C0,
  T0,
  shape_alpha,
  rate_alpha,
  X,
  L20,
  Mu0,
  W1,
  W2,
  V1,
  V2,
  L1,
  L2,
  Plog,
  maxit
)
}
\arguments{
\item{d}{Number of Dimension}

\item{n}{Number of Observations}

\item{C0}{Covariance matrix of the Observed data}

\item{T0}{Number of clusters in the variational distribution}

\item{shape_alpha}{Shape parameter of the prior Gamma distribution for alpha}

\item{rate_alpha}{Rate parameter of the prior Gamma distribution for alpha}

\item{X}{Observations, n x d matrix}

\item{L20}{Precision parameter of the prior multivariate normal distribution for eta}

\item{Mu0}{Mean parameter of the prior multivariate normal distribution for eta}

\item{W1}{Shape parameter of the posterior gamma distribution for alpha}

\item{W2}{Rate parameter of the posterior gamma distribution for alpha}

\item{V1}{1st shape parameters of the posterior beta distribution for v's, T0 x 1 matrix}

\item{V2}{2nd shape parameters of the posterior beta distribution for v's, T0 x 1 matrix}

\item{L1}{Mean parameters of the posterior multivariate normal distribution for eta's, T0 x d matrix}

\item{L2}{Precision parameters of the posterior multivariate normal distribution for eta's, T0 x 1 matrix}

\item{Plog}{Posterior probabilities of the latent allocations}

\item{maxit}{Maximum number of iterations for the variational inference}
}
\value{
the posterior expected value of alpha and the number of clusters on the basis of latent allocation probabilities
}
\description{
This function performs the mean-field variational inference for a DP mixture model, where the alpha parameter from the DP prior is considered varied (a Gamma prior) and the variance of the observations is considered fixed
}
\examples{
mfvi1(d=2,n=100,C0=diag(2),T0=20,shape_alpha=0.001,rate_alpha=0.001,X=matrix(1,nrow=100, ncol=2),L20=0.001,Mu0=matrix(c(rep(0,2)), nrow=1),W1=0.001,W2=0.001,V1=matrix(0.0001, nrow = 20, ncol = 1),V2=matrix(0.0001, nrow = 20, ncol = 1),L1=matrix(0.0001, nrow = 20, ncol = 2),L2=matrix(0.0001, nrow = 20, ncol = 1),Plog=matrix(-3,nrow=100,ncol=20),maxit=1000)
}
